~moved `expression.ts`
